requirements.txt
tensorflow==2.11.0
Pillow==9.0.0  # For image processing
numpy<2

